Legacy/requirements_20191129.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
asn1crypto=1.2.0=py37_0
astroid=2.3.3=py37_0
blas=1.0=mkl
ca-certificates=2019.10.16=0
certifi=2019.9.11=py37_0
cffi=1.13.2=py37h2e261b9_0
chardet=3.0.4=py37_1003
cryptography=2.8=py37h1ba5d50_0
cyrus-sasl=2.1.26=h82bf5a1_4
icu=58.2=h9c2bf20_1
idna=2.8=py37_0
intel-openmp=2019.4=243
isort=4.3.21=py37_0
krb5=1.16.1=h173b8e3_7
lazy-object-proxy=1.4.3=py37h7b6447c_0
libdb=6.1.26=he6710b0_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libntlm=1.5=h7b6447c_0
libprotobuf=3.6.0=hdbcaa40_0
libstdcxx-ng=9.1.0=hdf63c60_0
libxml2=2.9.9=hea5a465_1
libxslt=1.1.33=h7d1a2b0_0
lxml=4.4.1=py37hefd8a0e_0
mccabe=0.6.1=py37_1
mkl=2019.4=243
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
mysql=5.7.24=hbb652a2_0
mysql-connector-c=6.1.11=h597af5e_0
mysql-connector-python=8.0.18=py37h9c95fcb_1
ncurses=6.1=he6710b0_1
numpy=1.17.3=py37hd14ec0e_0
numpy-base=1.17.3=py37hde5b4d6_0
openssl=1.1.1d=h7b6447c_3
pandas=0.25.3=py37he6710b0_0
pip=19.3.1=py37_0
protobuf=3.6.0=py37hf484d3e_0
pycparser=2.19=py_0
pylint=2.4.4=py37_0
pyopenssl=19.1.0=py37_0
pysocks=1.7.1=py37_0
python=3.7.5=h0371630_0
python-dateutil=2.8.1=py_0
pytz=2019.3=py_0
readline=7.0=h7b6447c_5
requests=2.22.0=py37_0
setuptools=41.6.0=py37_0
six=1.13.0=py37_0
sqlite=3.30.1=h7b6447c_0
tk=8.6.8=hbc83047_0
urllib3=1.24.2=py37_0
wheel=0.33.6=py37_0
wrapt=1.11.2=py37h7b6447c_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
